requirements.txt
# Mistral & AI
mistralai
huggingface_hub
python-dotenv

# F1 & Data
fastf1
pandas
numpy

# Vector Search
faiss-cpu

# Vision & Processing
opencv-python
pillow
PyMuPDF

# Web/API (Optional for UI)
fastapi
uvicorn